Fix critical crash bugs + harden RTDB security rules across all nodes
- User.kt: accuracy Float->Double to fix Firebase type mismatch crash
- FirebaseRepository.kt: try-catch in listenToChildUsers() to prevent corrupt node crash
- MyFirebaseMessagingService.kt: startService->startForegroundService for Android 8+ compat
- database.rules.json: per-childUid isolation for all 15 sensitive data nodes (call_logs, sms_logs, location_history, alerts, notifications, schedules, snapshots, app_usage, commands, geofences, web_history, network_history, sim_info, package_events, media_items)

+++ app/src/main/java/com/example/authapp/data/User.kt
@@ -7,7 +7,7 @@ import com.google.firebase.database.PropertyName
 data class UserLocation(
     val latitude: Double = 0.0,
     val longitude: Double = 0.0,
-    val accuracy: Float = 0f,
+    val accuracy: Double = 0.0,
     val timestamp: Long = 0L,
     val provider: String = ""
 )
